Change default terminal to kitty
Alacritty had an issue w/ colors, where certain colors in neovim would
be a strange color when the following was true:

- transparent bg in alacritty
- using tmux

This was a minor frustrationg, but it was still a frustration, and I
didn't have meaningful reason to use alacritty instead. Kitty doesn't
have that problem (and seems a little bit more mature, both in terms of
the project's status and the behavior of its maintainers) so for now at
least we're switching.